+++ DoAnCuoiKy/DoAnCuoiKy/Model/Request/BookRequest.cs
@@ -13,7 +13,7 @@ namespace DoAnCuoiKy.Model.Request
         public int? YearPublished { get; set; }
         public int Quantity { get; set; } //số lượng sách nhập
         //public float? TotalPrice { get; set; }
-        public float? UnitPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
         public Guid? CategoryId { get; set; }
        
         public Guid? BookChapterId { get; set; }
