+++ DoAnCuoiKy/DoAnCuoiKy/Model/Request/FineRequest.cs
@@ -5,6 +5,6 @@ namespace DoAnCuoiKy.Model.Request
     public class FineRequest
     {
         public FineReason? fineReason { get; set; }
-        public float FineRate { get; set; }
+        public decimal FineRate { get; set; }
     }
 }
