+++ DoAnCuoiKy/DoAnCuoiKy/Model/Response/BookResponse.cs
@@ -14,8 +14,8 @@ namespace DoAnCuoiKy.Model.Response
         public string? PublisherName { get; set; }
         public int? YearPublished { get; set; }
         public int? Quantity { get; set; } //số lượng sách nhập vào
-        public float? TotalPrice { get; set; }
-        public float? UnitPrice { get; set; }
+        public decimal? TotalPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
         public string? CategoryName { get; set; }
         public string? TitleBookChapter { get; set; }
         public string Description { get; set; }
