+++ DoAnCuoiKy/DoAnCuoiKy/Model/Response/BorrowingDetailForFineResponse.cs
@@ -6,7 +6,7 @@ namespace DoAnCuoiKy.Model.Response
     {
         public Guid BorrowingDetailId { get; set; }
         public string BookTitle { get; set; }
-        public float UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public List<FineResponse> fineResponses { get; set; }
         public DateTime ReturnedDate { get; set; }
         public bool? IsFined { get; set; }
